// packages/dnb-eufemia/src/components/modal/types.d.ts
export type CloseHandlerParams = {
    triggeredBy: TriggeredBy;
    triggeredByEvent?: Event;
};
export type CloseHandler = (params?: CloseHandlerParams) => void;
export interface ModalProps extends ModalRootProps {
    /**
     * The id used internal in the modal/drawer root element. Defaults to `root`, so the element id will be `dnb-modal-root`.
     */
    id?: string;
    /**
     * Will disable the trigger button.
     */
    disabled?: boolean;
    /**
     * Forces the modal/drawer to delay the opening. The delay is given in `ms`.
     */
    openDelay?: string | number;
    /**
     * If set to `true` (boolean or string), then the user can&#39;t close the modal/drawer.
     */
    preventClose?: boolean;
    /**
     * Duration of animation open/close in ms. Defaults to 300ms.
     */
    animationDuration?: string | number;
    /**
     * If set to `true`, no open/close animation will be shown. Defaults to false.
     */
    noAnimation?: boolean;
    /**
     * Use this prop to control the open/close state by setting `true` / `false`.
     */
    open?: boolean;
    /**
     * The content which will appear when triggering the modal/drawer.
     */
    children?: React.ReactNode | ((props: ModalProps) => React.ReactNode);
    /**
     * Omits default showing trigger button
     */
    omitTriggerButton?: boolean;
    /**
     * This event gets triggered once the modal shows up. Returns the modal id: `{ id }`.
     */
    onOpen?: ({ id }: {
        id?: string;
    }) => void;
    /**
     * This event gets triggered once the modal gets closed. Returns the modal id: `{ id, event, triggeredBy }`.
     */
    onClose?: ({ id, event, triggeredBy, }: {
        id?: string;
        event?: Event;
        triggeredBy?: string;
    }) => void;
    /**
     * This event gets triggered once the user tries to close the modal, but `preventClose` is set to "true". Returns a callback `close` you can call to trigger the close mechanism. More details below. Returns the modal id: `{ id, event, close: Method, triggeredBy }`
     */
    onClosePrevent?: ({ id, event, triggeredBy, close, }: {
        id?: string;
        event?: Event;
        triggeredBy?: string;
        close?: CloseHandler;
    }) => void;
    /**
     * Set a function to call the callback function, once the modal/drawer should open: `openModal={(open) => open()}`
     */
    openModal?: (open?: (e: Event) => void, instance?: any) => () => void | void;
    /**
     * Set a function to call the callback function, once the modal/drawer should close: `closeModal={(close) => close()}`
     */
    closeModal?: (close?: CloseHandler, instance?: any) => () => void | void;
    /**
     * Provide a custom trigger component. Like trigger={<Anchor href="/" />}. It will set the focus on it when the modal/drawer gets closed.
     */
    trigger?: React.ElementType;
    /**
     * Send along custom HTML attributes or properties to the trigger button.
     */
    triggerAttributes?: TriggerAttributes;
    /**
     * The content which will appear when triggering the modal/drawer.
     */
    modalContent?: React.ReactNode | ((props: ModalProps) => React.ReactNode);
    /**
     * If true, the drawer will not open in a new DOM but directly in current DOM. Defaults to `false`.
     */
    directDomReturn?: boolean;
    /**
     * To get the inner content Element, pass in your own React ref
     */
    contentRef?: React.RefObject<HTMLElement>;
    /**
     * To get the scroll Element, pass in your own React ref
     */
    scrollRef?: React.RefObject<HTMLElement>;
}
// ...
export type TriggerAttributes = ButtonProps;